test(crates): split the proof tests by proof kind
Move the compliance, logic, and aggregation proof tests from
`proofs.rs` into one file per proof kind, with shared helpers in
`tests/common`.
